[HUDI-9152] Improve read/write/compaction performance by reusing avro schema (#12949)
1. Introduce JVM level caching for avro schema to reduce the cost of schema comparison.

NOTE: Use cache to cache references to the schema on key links where the schema may be created repeatedly.
This ensures that only one variable instance of the same Schema will be used during a JVM lifetime,
thus reducing the overhead of schema comparison on important io paths. For most of the cases, we only need to compare whether it is the same reference, there is no need to call the `Schema::equals` method.

2. Cache the frequently reused Schema on the IO code path.

---------

Signed-off-by: TheR1sing3un <chaoyang@apache.org>